Move listeners back to their own project
